A separate folder was opened for classes. The methods in the class were moved here and deleted from program.cs. Then, the class in the folder was inherited to the program class.
