+++ Contructor/Program.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Contructor.Classes;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,35 +7,21 @@ using System.Threading.Tasks;
 
 namespace Contructor
 {
-    public class Program
+    public class Program : CustomerManager
     {
+        public Program(int count) : base(count)
+        {
+        }
+
         static void Main(string[] args)
         {
             CustomerManager customerManager = new CustomerManager(0129347 );
             customerManager.List();
+            customerManager.Add();
 
             Console.ReadLine();
 
 
-
-        }
-
-    }
-    class CustomerManager
-    {
-        private int _count;
-        public CustomerManager(int count)
-        {
-                _count = count;
-        }
-
-        public void List()
-        {
-            Console.WriteLine("Listed {0}", _count);
-        }
-        public void Add() 
-        {
-            Console.WriteLine("Added");
         }
     }
